+++ .storybook/main.js
@@ -1,7 +1,16 @@
+const {
+  dirname,
+  join
+} = require("node:path");
+
 module.exports = {
   stories: [
     '../packages/decap-cms-core/src/**/*.stories.js',
     '../packages/decap-cms-ui-default/src/**/*.stories.js',
   ],
-  addons: ['@storybook/addon-actions', '@storybook/addon-links'],
+  addons: [getAbsolutePath("@storybook/addon-links")],
 };
+
+function getAbsolutePath(value) {
+  return dirname(require.resolve(join(value, "package.json")));
+}