+++ packages/decap-cms-backend-bitbucket/src/AuthenticationPage.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { NetlifyAuthenticator, ImplicitAuthenticator } from 'decap-cms-lib-auth';
@@ -8,7 +8,7 @@ const LoginButtonIcon = styled(Icon)`
   margin-right: 18px;
 `;
 
-export default class BitbucketAuthenticationPage extends React.Component {
+export default class BitbucketAuthenticationPage extends Component {
   static propTypes = {
     onLogin: PropTypes.func.isRequired,
     inProgress: PropTypes.bool,
@@ -91,10 +91,10 @@ export default class BitbucketAuthenticationPage extends React.Component {
         logo={config.logo}
         siteUrl={config.site_url}
         renderButtonContent={() => (
-          <React.Fragment>
+          <Fragment>
             <LoginButtonIcon type="bitbucket" />
             {inProgress ? t('auth.loggingIn') : t('auth.loginWithBitbucket')}
-          </React.Fragment>
+          </Fragment>
         )}
         t={t}
       />