+++ packages/decap-cms/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.16.0](https://github.com/decaporg/decap-cms/compare/decap-cms@3.15.1...decap-cms@3.16.0) (2026-08-31)
+
+**Note:** Version bump only for package decap-cms
+
 ## [3.15.1](https://github.com/decaporg/decap-cms/compare/decap-cms@3.15.0...decap-cms@3.15.1) (2026-07-24)
 
 **Note:** Version bump only for package decap-cms